+++ packages/core/src/__tests__/conversation-action-examples.test.ts
@@ -12,7 +12,7 @@ import {
   createConversationActions,
   type ConversationActionDeps,
 } from "../conversation-actions.js";
-import { createConvosOnboardingScheme } from "../convos/onboarding-scheme.js";
+import { createConvosOnboardingScheme } from "../schemes/convos/onboarding-scheme.js";
 
 function stubCtx(): HandlerContext {
   return {
