+++ packages/core/src/__tests__/sdk-client.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, test } from "bun:test";
-import { encodeProfileSnapshot } from "../convos/profile-messages.js";
-import { createConvosOnboardingScheme } from "../convos/onboarding-scheme.js";
+import { encodeProfileSnapshot } from "../schemes/convos/profile-messages.js";
+import { createConvosOnboardingScheme } from "../schemes/convos/onboarding-scheme.js";
 import type { EncodedOnboardingContent } from "../schemes/onboarding-scheme.js";
 import { createSdkClient } from "../sdk/sdk-client.js";
 import {
