+++ packages/core/src/schemes/convos/__tests__/invite-join-error.test.ts
