+++ packages/core/src/schemes/convos/__tests__/join-request-content.test.ts
