+++ packages/core/src/schemes/convos/invite-join-error.ts
