+++ packages/core/src/schemes/convos/join-request-content.ts
