+++ packages/core/src/schemes/join.ts
@@ -2,5 +2,5 @@ export type {
   JoinConversationDeps,
   JoinConversationOptions,
   JoinResult,
-} from "../convos/join.js";
-export { joinConversation } from "../convos/join.js";
+} from "./convos/join.js";
+export { joinConversation } from "./convos/join.js";
